% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boolean_input_mult_df}
\alias{boolean_input_mult_df}
\title{Sparse Boolean dataframe with rare variant information
and multiple outcome variables}
\format{
A data frame with 5000 rows and 1004 variables:
\describe{
 \item{Sample_Name}{Unique identifier of the samples}
 \item{Input_1}{Presence and absense of rare variant 1}
 \item{Input_2}{Presence and absense of rare variant 2}
 \item{Input_3}{Presence and absense of rare variant 3}
 \item{Input_4}{Presence and absense of rare variant 4}
 \item{Input_5}{Presence and absense of rare variant 5}
 \item{Input_6}{Presence and absense of rare variant 6}
 \item{Input_7}{Presence and absense of rare variant 7}
 \item{Input_8}{Presence and absense of rare variant 8}
 \item{Input_9}{Presence and absense of rare variant 9}
 \item{Input_10}{Presence and absense of rare variant 10}
 \item{Input_11}{Presence and absense of rare variant 11}
 \item{Input_12}{Presence and absense of rare variant 12}
 \item{Input_13}{Presence and absense of rare variant 13}
 \item{Input_14}{Presence and absense of rare variant 14}
 \item{Input_15}{Presence and absense of rare variant 15}
 \item{Input_16}{Presence and absense of rare variant 16}
 \item{Input_17}{Presence and absense of rare variant 17}
 \item{Input_18}{Presence and absense of rare variant 18}
 \item{Input_19}{Presence and absense of rare variant 19}
 \item{Input_20}{Presence and absense of rare variant 20}
 \item{Input_21}{Presence and absense of rare variant 21}
 \item{Input_22}{Presence and absense of rare variant 22}
 \item{Input_23}{Presence and absense of rare variant 23}
 \item{Input_24}{Presence and absense of rare variant 24}
 \item{Input_25}{Presence and absense of rare variant 25}
 \item{Input_26}{Presence and absense of rare variant 26}
 \item{Input_27}{Presence and absense of rare variant 27}
 \item{Input_28}{Presence and absense of rare variant 28}
 \item{Input_29}{Presence and absense of rare variant 29}
 \item{Input_30}{Presence and absense of rare variant 30}
 \item{Input_31}{Presence and absense of rare variant 31}
 \item{Input_32}{Presence and absense of rare variant 32}
 \item{Input_33}{Presence and absense of rare variant 33}
 \item{Input_34}{Presence and absense of rare variant 34}
 \item{Input_35}{Presence and absense of rare variant 35}
 \item{Input_36}{Presence and absense of rare variant 36}
 \item{Input_37}{Presence and absense of rare variant 37}
 \item{Input_38}{Presence and absense of rare variant 38}
 \item{Input_39}{Presence and absense of rare variant 39}
 \item{Input_40}{Presence and absense of rare variant 40}
 \item{Input_41}{Presence and absense of rare variant 41}
 \item{Input_42}{Presence and absense of rare variant 42}
 \item{Input_43}{Presence and absense of rare variant 43}
 \item{Input_44}{Presence and absense of rare variant 44}
 \item{Input_45}{Presence and absense of rare variant 45}
 \item{Input_46}{Presence and absense of rare variant 46}
 \item{Input_47}{Presence and absense of rare variant 47}
 \item{Input_48}{Presence and absense of rare variant 48}
 \item{Input_49}{Presence and absense of rare variant 49}
 \item{Input_50}{Presence and absense of rare variant 50}
 \item{Input_51}{Presence and absense of rare variant 51}
 \item{Input_52}{Presence and absense of rare variant 52}
 \item{Input_53}{Presence and absense of rare variant 53}
 \item{Input_54}{Presence and absense of rare variant 54}
 \item{Input_55}{Presence and absense of rare variant 55}
 \item{Input_56}{Presence and absense of rare variant 56}
 \item{Input_57}{Presence and absense of rare variant 57}
 \item{Input_58}{Presence and absense of rare variant 58}
 \item{Input_59}{Presence and absense of rare variant 59}
 \item{Input_60}{Presence and absense of rare variant 60}
 \item{Input_61}{Presence and absense of rare variant 61}
 \item{Input_62}{Presence and absense of rare variant 62}
 \item{Input_63}{Presence and absense of rare variant 63}
 \item{Input_64}{Presence and absense of rare variant 64}
 \item{Input_65}{Presence and absense of rare variant 65}
 \item{Input_66}{Presence and absense of rare variant 66}
 \item{Input_67}{Presence and absense of rare variant 67}
 \item{Input_68}{Presence and absense of rare variant 68}
 \item{Input_69}{Presence and absense of rare variant 69}
 \item{Input_70}{Presence and absense of rare variant 70}
 \item{Input_71}{Presence and absense of rare variant 71}
 \item{Input_72}{Presence and absense of rare variant 72}
 \item{Input_73}{Presence and absense of rare variant 73}
 \item{Input_74}{Presence and absense of rare variant 74}
 \item{Input_75}{Presence and absense of rare variant 75}
 \item{Input_76}{Presence and absense of rare variant 76}
 \item{Input_77}{Presence and absense of rare variant 77}
 \item{Input_78}{Presence and absense of rare variant 78}
 \item{Input_79}{Presence and absense of rare variant 79}
 \item{Input_80}{Presence and absense of rare variant 80}
 \item{Input_81}{Presence and absense of rare variant 81}
 \item{Input_82}{Presence and absense of rare variant 82}
 \item{Input_83}{Presence and absense of rare variant 83}
 \item{Input_84}{Presence and absense of rare variant 84}
 \item{Input_85}{Presence and absense of rare variant 85}
 \item{Input_86}{Presence and absense of rare variant 86}
 \item{Input_87}{Presence and absense of rare variant 87}
 \item{Input_88}{Presence and absense of rare variant 88}
 \item{Input_89}{Presence and absense of rare variant 89}
 \item{Input_90}{Presence and absense of rare variant 90}
 \item{Input_91}{Presence and absense of rare variant 91}
 \item{Input_92}{Presence and absense of rare variant 92}
 \item{Input_93}{Presence and absense of rare variant 93}
 \item{Input_94}{Presence and absense of rare variant 94}
 \item{Input_95}{Presence and absense of rare variant 95}
 \item{Input_96}{Presence and absense of rare variant 96}
 \item{Input_97}{Presence and absense of rare variant 97}
 \item{Input_98}{Presence and absense of rare variant 98}
 \item{Input_99}{Presence and absense of rare variant 99}
 \item{Input_100}{Presence and absense of rare variant 100}
 \item{Input_101}{Presence and absense of rare variant 101}
 \item{Input_102}{Presence and absense of rare variant 102}
 \item{Input_103}{Presence and absense of rare variant 103}
 \item{Input_104}{Presence and absense of rare variant 104}
 \item{Input_105}{Presence and absense of rare variant 105}
 \item{Input_106}{Presence and absense of rare variant 106}
 \item{Input_107}{Presence and absense of rare variant 107}
 \item{Input_108}{Presence and absense of rare variant 108}
 \item{Input_109}{Presence and absense of rare variant 109}
 \item{Input_110}{Presence and absense of rare variant 110}
 \item{Input_111}{Presence and absense of rare variant 111}
 \item{Input_112}{Presence and absense of rare variant 112}
 \item{Input_113}{Presence and absense of rare variant 113}
 \item{Input_114}{Presence and absense of rare variant 114}
 \item{Input_115}{Presence and absense of rare variant 115}
 \item{Input_116}{Presence and absense of rare variant 116}
 \item{Input_117}{Presence and absense of rare variant 117}
 \item{Input_118}{Presence and absense of rare variant 118}
 \item{Input_119}{Presence and absense of rare variant 119}
 \item{Input_120}{Presence and absense of rare variant 120}
 \item{Input_121}{Presence and absense of rare variant 121}
 \item{Input_122}{Presence and absense of rare variant 122}
 \item{Input_123}{Presence and absense of rare variant 123}
 \item{Input_124}{Presence and absense of rare variant 124}
 \item{Input_125}{Presence and absense of rare variant 125}
 \item{Input_126}{Presence and absense of rare variant 126}
 \item{Input_127}{Presence and absense of rare variant 127}
 \item{Input_128}{Presence and absense of rare variant 128}
 \item{Input_129}{Presence and absense of rare variant 129}
 \item{Input_130}{Presence and absense of rare variant 130}
 \item{Input_131}{Presence and absense of rare variant 131}
 \item{Input_132}{Presence and absense of rare variant 132}
 \item{Input_133}{Presence and absense of rare variant 133}
 \item{Input_134}{Presence and absense of rare variant 134}
 \item{Input_135}{Presence and absense of rare variant 135}
 \item{Input_136}{Presence and absense of rare variant 136}
 \item{Input_137}{Presence and absense of rare variant 137}
 \item{Input_138}{Presence and absense of rare variant 138}
 \item{Input_139}{Presence and absense of rare variant 139}
 \item{Input_140}{Presence and absense of rare variant 140}
 \item{Input_141}{Presence and absense of rare variant 141}
 \item{Input_142}{Presence and absense of rare variant 142}
 \item{Input_143}{Presence and absense of rare variant 143}
 \item{Input_144}{Presence and absense of rare variant 144}
 \item{Input_145}{Presence and absense of rare variant 145}
 \item{Input_146}{Presence and absense of rare variant 146}
 \item{Input_147}{Presence and absense of rare variant 147}
 \item{Input_148}{Presence and absense of rare variant 148}
 \item{Input_149}{Presence and absense of rare variant 149}
 \item{Input_150}{Presence and absense of rare variant 150}
 \item{Input_151}{Presence and absense of rare variant 151}
 \item{Input_152}{Presence and absense of rare variant 152}
 \item{Input_153}{Presence and absense of rare variant 153}
 \item{Input_154}{Presence and absense of rare variant 154}
 \item{Input_155}{Presence and absense of rare variant 155}
 \item{Input_156}{Presence and absense of rare variant 156}
 \item{Input_157}{Presence and absense of rare variant 157}
 \item{Input_158}{Presence and absense of rare variant 158}
 \item{Input_159}{Presence and absense of rare variant 159}
 \item{Input_160}{Presence and absense of rare variant 160}
 \item{Input_161}{Presence and absense of rare variant 161}
 \item{Input_162}{Presence and absense of rare variant 162}
 \item{Input_163}{Presence and absense of rare variant 163}
 \item{Input_164}{Presence and absense of rare variant 164}
 \item{Input_165}{Presence and absense of rare variant 165}
 \item{Input_166}{Presence and absense of rare variant 166}
 \item{Input_167}{Presence and absense of rare variant 167}
 \item{Input_168}{Presence and absense of rare variant 168}
 \item{Input_169}{Presence and absense of rare variant 169}
 \item{Input_170}{Presence and absense of rare variant 170}
 \item{Input_171}{Presence and absense of rare variant 171}
 \item{Input_172}{Presence and absense of rare variant 172}
 \item{Input_173}{Presence and absense of rare variant 173}
 \item{Input_174}{Presence and absense of rare variant 174}
 \item{Input_175}{Presence and absense of rare variant 175}
 \item{Input_176}{Presence and absense of rare variant 176}
 \item{Input_177}{Presence and absense of rare variant 177}
 \item{Input_178}{Presence and absense of rare variant 178}
 \item{Input_179}{Presence and absense of rare variant 179}
 \item{Input_180}{Presence and absense of rare variant 180}
 \item{Input_181}{Presence and absense of rare variant 181}
 \item{Input_182}{Presence and absense of rare variant 182}
 \item{Input_183}{Presence and absense of rare variant 183}
 \item{Input_184}{Presence and absense of rare variant 184}
 \item{Input_185}{Presence and absense of rare variant 185}
 \item{Input_186}{Presence and absense of rare variant 186}
 \item{Input_187}{Presence and absense of rare variant 187}
 \item{Input_188}{Presence and absense of rare variant 188}
 \item{Input_189}{Presence and absense of rare variant 189}
 \item{Input_190}{Presence and absense of rare variant 190}
 \item{Input_191}{Presence and absense of rare variant 191}
 \item{Input_192}{Presence and absense of rare variant 192}
 \item{Input_193}{Presence and absense of rare variant 193}
 \item{Input_194}{Presence and absense of rare variant 194}
 \item{Input_195}{Presence and absense of rare variant 195}
 \item{Input_196}{Presence and absense of rare variant 196}
 \item{Input_197}{Presence and absense of rare variant 197}
 \item{Input_198}{Presence and absense of rare variant 198}
 \item{Input_199}{Presence and absense of rare variant 199}
 \item{Input_200}{Presence and absense of rare variant 200}
 \item{Input_201}{Presence and absense of rare variant 201}
 \item{Input_202}{Presence and absense of rare variant 202}
 \item{Input_203}{Presence and absense of rare variant 203}
 \item{Input_204}{Presence and absense of rare variant 204}
 \item{Input_205}{Presence and absense of rare variant 205}
 \item{Input_206}{Presence and absense of rare variant 206}
 \item{Input_207}{Presence and absense of rare variant 207}
 \item{Input_208}{Presence and absense of rare variant 208}
 \item{Input_209}{Presence and absense of rare variant 209}
 \item{Input_210}{Presence and absense of rare variant 210}
 \item{Input_211}{Presence and absense of rare variant 211}
 \item{Input_212}{Presence and absense of rare variant 212}
 \item{Input_213}{Presence and absense of rare variant 213}
 \item{Input_214}{Presence and absense of rare variant 214}
 \item{Input_215}{Presence and absense of rare variant 215}
 \item{Input_216}{Presence and absense of rare variant 216}
 \item{Input_217}{Presence and absense of rare variant 217}
 \item{Input_218}{Presence and absense of rare variant 218}
 \item{Input_219}{Presence and absense of rare variant 219}
 \item{Input_220}{Presence and absense of rare variant 220}
 \item{Input_221}{Presence and absense of rare variant 221}
 \item{Input_222}{Presence and absense of rare variant 222}
 \item{Input_223}{Presence and absense of rare variant 223}
 \item{Input_224}{Presence and absense of rare variant 224}
 \item{Input_225}{Presence and absense of rare variant 225}
 \item{Input_226}{Presence and absense of rare variant 226}
 \item{Input_227}{Presence and absense of rare variant 227}
 \item{Input_228}{Presence and absense of rare variant 228}
 \item{Input_229}{Presence and absense of rare variant 229}
 \item{Input_230}{Presence and absense of rare variant 230}
 \item{Input_231}{Presence and absense of rare variant 231}
 \item{Input_232}{Presence and absense of rare variant 232}
 \item{Input_233}{Presence and absense of rare variant 233}
 \item{Input_234}{Presence and absense of rare variant 234}
 \item{Input_235}{Presence and absense of rare variant 235}
 \item{Input_236}{Presence and absense of rare variant 236}
 \item{Input_237}{Presence and absense of rare variant 237}
 \item{Input_238}{Presence and absense of rare variant 238}
 \item{Input_239}{Presence and absense of rare variant 239}
 \item{Input_240}{Presence and absense of rare variant 240}
 \item{Input_241}{Presence and absense of rare variant 241}
 \item{Input_242}{Presence and absense of rare variant 242}
 \item{Input_243}{Presence and absense of rare variant 243}
 \item{Input_244}{Presence and absense of rare variant 244}
 \item{Input_245}{Presence and absense of rare variant 245}
 \item{Input_246}{Presence and absense of rare variant 246}
 \item{Input_247}{Presence and absense of rare variant 247}
 \item{Input_248}{Presence and absense of rare variant 248}
 \item{Input_249}{Presence and absense of rare variant 249}
 \item{Input_250}{Presence and absense of rare variant 250}
 \item{Input_251}{Presence and absense of rare variant 251}
 \item{Input_252}{Presence and absense of rare variant 252}
 \item{Input_253}{Presence and absense of rare variant 253}
 \item{Input_254}{Presence and absense of rare variant 254}
 \item{Input_255}{Presence and absense of rare variant 255}
 \item{Input_256}{Presence and absense of rare variant 256}
 \item{Input_257}{Presence and absense of rare variant 257}
 \item{Input_258}{Presence and absense of rare variant 258}
 \item{Input_259}{Presence and absense of rare variant 259}
 \item{Input_260}{Presence and absense of rare variant 260}
 \item{Input_261}{Presence and absense of rare variant 261}
 \item{Input_262}{Presence and absense of rare variant 262}
 \item{Input_263}{Presence and absense of rare variant 263}
 \item{Input_264}{Presence and absense of rare variant 264}
 \item{Input_265}{Presence and absense of rare variant 265}
 \item{Input_266}{Presence and absense of rare variant 266}
 \item{Input_267}{Presence and absense of rare variant 267}
 \item{Input_268}{Presence and absense of rare variant 268}
 \item{Input_269}{Presence and absense of rare variant 269}
 \item{Input_270}{Presence and absense of rare variant 270}
 \item{Input_271}{Presence and absense of rare variant 271}
 \item{Input_272}{Presence and absense of rare variant 272}
 \item{Input_273}{Presence and absense of rare variant 273}
 \item{Input_274}{Presence and absense of rare variant 274}
 \item{Input_275}{Presence and absense of rare variant 275}
 \item{Input_276}{Presence and absense of rare variant 276}
 \item{Input_277}{Presence and absense of rare variant 277}
 \item{Input_278}{Presence and absense of rare variant 278}
 \item{Input_279}{Presence and absense of rare variant 279}
 \item{Input_280}{Presence and absense of rare variant 280}
 \item{Input_281}{Presence and absense of rare variant 281}
 \item{Input_282}{Presence and absense of rare variant 282}
 \item{Input_283}{Presence and absense of rare variant 283}
 \item{Input_284}{Presence and absense of rare variant 284}
 \item{Input_285}{Presence and absense of rare variant 285}
 \item{Input_286}{Presence and absense of rare variant 286}
 \item{Input_287}{Presence and absense of rare variant 287}
 \item{Input_288}{Presence and absense of rare variant 288}
 \item{Input_289}{Presence and absense of rare variant 289}
 \item{Input_290}{Presence and absense of rare variant 290}
 \item{Input_291}{Presence and absense of rare variant 291}
 \item{Input_292}{Presence and absense of rare variant 292}
 \item{Input_293}{Presence and absense of rare variant 293}
 \item{Input_294}{Presence and absense of rare variant 294}
 \item{Input_295}{Presence and absense of rare variant 295}
 \item{Input_296}{Presence and absense of rare variant 296}
 \item{Input_297}{Presence and absense of rare variant 297}
 \item{Input_298}{Presence and absense of rare variant 298}
 \item{Input_299}{Presence and absense of rare variant 299}
 \item{Input_300}{Presence and absense of rare variant 300}
 \item{Input_301}{Presence and absense of rare variant 301}
 \item{Input_302}{Presence and absense of rare variant 302}
 \item{Input_303}{Presence and absense of rare variant 303}
 \item{Input_304}{Presence and absense of rare variant 304}
 \item{Input_305}{Presence and absense of rare variant 305}
 \item{Input_306}{Presence and absense of rare variant 306}
 \item{Input_307}{Presence and absense of rare variant 307}
 \item{Input_308}{Presence and absense of rare variant 308}
 \item{Input_309}{Presence and absense of rare variant 309}
 \item{Input_310}{Presence and absense of rare variant 310}
 \item{Input_311}{Presence and absense of rare variant 311}
 \item{Input_312}{Presence and absense of rare variant 312}
 \item{Input_313}{Presence and absense of rare variant 313}
 \item{Input_314}{Presence and absense of rare variant 314}
 \item{Input_315}{Presence and absense of rare variant 315}
 \item{Input_316}{Presence and absense of rare variant 316}
 \item{Input_317}{Presence and absense of rare variant 317}
 \item{Input_318}{Presence and absense of rare variant 318}
 \item{Input_319}{Presence and absense of rare variant 319}
 \item{Input_320}{Presence and absense of rare variant 320}
 \item{Input_321}{Presence and absense of rare variant 321}
 \item{Input_322}{Presence and absense of rare variant 322}
 \item{Input_323}{Presence and absense of rare variant 323}
 \item{Input_324}{Presence and absense of rare variant 324}
 \item{Input_325}{Presence and absense of rare variant 325}
 \item{Input_326}{Presence and absense of rare variant 326}
 \item{Input_327}{Presence and absense of rare variant 327}
 \item{Input_328}{Presence and absense of rare variant 328}
 \item{Input_329}{Presence and absense of rare variant 329}
 \item{Input_330}{Presence and absense of rare variant 330}
 \item{Input_331}{Presence and absense of rare variant 331}
 \item{Input_332}{Presence and absense of rare variant 332}
 \item{Input_333}{Presence and absense of rare variant 333}
 \item{Input_334}{Presence and absense of rare variant 334}
 \item{Input_335}{Presence and absense of rare variant 335}
 \item{Input_336}{Presence and absense of rare variant 336}
 \item{Input_337}{Presence and absense of rare variant 337}
 \item{Input_338}{Presence and absense of rare variant 338}
 \item{Input_339}{Presence and absense of rare variant 339}
 \item{Input_340}{Presence and absense of rare variant 340}
 \item{Input_341}{Presence and absense of rare variant 341}
 \item{Input_342}{Presence and absense of rare variant 342}
 \item{Input_343}{Presence and absense of rare variant 343}
 \item{Input_344}{Presence and absense of rare variant 344}
 \item{Input_345}{Presence and absense of rare variant 345}
 \item{Input_346}{Presence and absense of rare variant 346}
 \item{Input_347}{Presence and absense of rare variant 347}
 \item{Input_348}{Presence and absense of rare variant 348}
 \item{Input_349}{Presence and absense of rare variant 349}
 \item{Input_350}{Presence and absense of rare variant 350}
 \item{Input_351}{Presence and absense of rare variant 351}
 \item{Input_352}{Presence and absense of rare variant 352}
 \item{Input_353}{Presence and absense of rare variant 353}
 \item{Input_354}{Presence and absense of rare variant 354}
 \item{Input_355}{Presence and absense of rare variant 355}
 \item{Input_356}{Presence and absense of rare variant 356}
 \item{Input_357}{Presence and absense of rare variant 357}
 \item{Input_358}{Presence and absense of rare variant 358}
 \item{Input_359}{Presence and absense of rare variant 359}
 \item{Input_360}{Presence and absense of rare variant 360}
 \item{Input_361}{Presence and absense of rare variant 361}
 \item{Input_362}{Presence and absense of rare variant 362}
 \item{Input_363}{Presence and absense of rare variant 363}
 \item{Input_364}{Presence and absense of rare variant 364}
 \item{Input_365}{Presence and absense of rare variant 365}
 \item{Input_366}{Presence and absense of rare variant 366}
 \item{Input_367}{Presence and absense of rare variant 367}
 \item{Input_368}{Presence and absense of rare variant 368}
 \item{Input_369}{Presence and absense of rare variant 369}
 \item{Input_370}{Presence and absense of rare variant 370}
 \item{Input_371}{Presence and absense of rare variant 371}
 \item{Input_372}{Presence and absense of rare variant 372}
 \item{Input_373}{Presence and absense of rare variant 373}
 \item{Input_374}{Presence and absense of rare variant 374}
 \item{Input_375}{Presence and absense of rare variant 375}
 \item{Input_376}{Presence and absense of rare variant 376}
 \item{Input_377}{Presence and absense of rare variant 377}
 \item{Input_378}{Presence and absense of rare variant 378}
 \item{Input_379}{Presence and absense of rare variant 379}
 \item{Input_380}{Presence and absense of rare variant 380}
 \item{Input_381}{Presence and absense of rare variant 381}
 \item{Input_382}{Presence and absense of rare variant 382}
 \item{Input_383}{Presence and absense of rare variant 383}
 \item{Input_384}{Presence and absense of rare variant 384}
 \item{Input_385}{Presence and absense of rare variant 385}
 \item{Input_386}{Presence and absense of rare variant 386}
 \item{Input_387}{Presence and absense of rare variant 387}
 \item{Input_388}{Presence and absense of rare variant 388}
 \item{Input_389}{Presence and absense of rare variant 389}
 \item{Input_390}{Presence and absense of rare variant 390}
 \item{Input_391}{Presence and absense of rare variant 391}
 \item{Input_392}{Presence and absense of rare variant 392}
 \item{Input_393}{Presence and absense of rare variant 393}
 \item{Input_394}{Presence and absense of rare variant 394}
 \item{Input_395}{Presence and absense of rare variant 395}
 \item{Input_396}{Presence and absense of rare variant 396}
 \item{Input_397}{Presence and absense of rare variant 397}
 \item{Input_398}{Presence and absense of rare variant 398}
 \item{Input_399}{Presence and absense of rare variant 399}
 \item{Input_400}{Presence and absense of rare variant 400}
 \item{Input_401}{Presence and absense of rare variant 401}
 \item{Input_402}{Presence and absense of rare variant 402}
 \item{Input_403}{Presence and absense of rare variant 403}
 \item{Input_404}{Presence and absense of rare variant 404}
 \item{Input_405}{Presence and absense of rare variant 405}
 \item{Input_406}{Presence and absense of rare variant 406}
 \item{Input_407}{Presence and absense of rare variant 407}
 \item{Input_408}{Presence and absense of rare variant 408}
 \item{Input_409}{Presence and absense of rare variant 409}
 \item{Input_410}{Presence and absense of rare variant 410}
 \item{Input_411}{Presence and absense of rare variant 411}
 \item{Input_412}{Presence and absense of rare variant 412}
 \item{Input_413}{Presence and absense of rare variant 413}
 \item{Input_414}{Presence and absense of rare variant 414}
 \item{Input_415}{Presence and absense of rare variant 415}
 \item{Input_416}{Presence and absense of rare variant 416}
 \item{Input_417}{Presence and absense of rare variant 417}
 \item{Input_418}{Presence and absense of rare variant 418}
 \item{Input_419}{Presence and absense of rare variant 419}
 \item{Input_420}{Presence and absense of rare variant 420}
 \item{Input_421}{Presence and absense of rare variant 421}
 \item{Input_422}{Presence and absense of rare variant 422}
 \item{Input_423}{Presence and absense of rare variant 423}
 \item{Input_424}{Presence and absense of rare variant 424}
 \item{Input_425}{Presence and absense of rare variant 425}
 \item{Input_426}{Presence and absense of rare variant 426}
 \item{Input_427}{Presence and absense of rare variant 427}
 \item{Input_428}{Presence and absense of rare variant 428}
 \item{Input_429}{Presence and absense of rare variant 429}
 \item{Input_430}{Presence and absense of rare variant 430}
 \item{Input_431}{Presence and absense of rare variant 431}
 \item{Input_432}{Presence and absense of rare variant 432}
 \item{Input_433}{Presence and absense of rare variant 433}
 \item{Input_434}{Presence and absense of rare variant 434}
 \item{Input_435}{Presence and absense of rare variant 435}
 \item{Input_436}{Presence and absense of rare variant 436}
 \item{Input_437}{Presence and absense of rare variant 437}
 \item{Input_438}{Presence and absense of rare variant 438}
 \item{Input_439}{Presence and absense of rare variant 439}
 \item{Input_440}{Presence and absense of rare variant 440}
 \item{Input_441}{Presence and absense of rare variant 441}
 \item{Input_442}{Presence and absense of rare variant 442}
 \item{Input_443}{Presence and absense of rare variant 443}
 \item{Input_444}{Presence and absense of rare variant 444}
 \item{Input_445}{Presence and absense of rare variant 445}
 \item{Input_446}{Presence and absense of rare variant 446}
 \item{Input_447}{Presence and absense of rare variant 447}
 \item{Input_448}{Presence and absense of rare variant 448}
 \item{Input_449}{Presence and absense of rare variant 449}
 \item{Input_450}{Presence and absense of rare variant 450}
 \item{Input_451}{Presence and absense of rare variant 451}
 \item{Input_452}{Presence and absense of rare variant 452}
 \item{Input_453}{Presence and absense of rare variant 453}
 \item{Input_454}{Presence and absense of rare variant 454}
 \item{Input_455}{Presence and absense of rare variant 455}
 \item{Input_456}{Presence and absense of rare variant 456}
 \item{Input_457}{Presence and absense of rare variant 457}
 \item{Input_458}{Presence and absense of rare variant 458}
 \item{Input_459}{Presence and absense of rare variant 459}
 \item{Input_460}{Presence and absense of rare variant 460}
 \item{Input_461}{Presence and absense of rare variant 461}
 \item{Input_462}{Presence and absense of rare variant 462}
 \item{Input_463}{Presence and absense of rare variant 463}
 \item{Input_464}{Presence and absense of rare variant 464}
 \item{Input_465}{Presence and absense of rare variant 465}
 \item{Input_466}{Presence and absense of rare variant 466}
 \item{Input_467}{Presence and absense of rare variant 467}
 \item{Input_468}{Presence and absense of rare variant 468}
 \item{Input_469}{Presence and absense of rare variant 469}
 \item{Input_470}{Presence and absense of rare variant 470}
 \item{Input_471}{Presence and absense of rare variant 471}
 \item{Input_472}{Presence and absense of rare variant 472}
 \item{Input_473}{Presence and absense of rare variant 473}
 \item{Input_474}{Presence and absense of rare variant 474}
 \item{Input_475}{Presence and absense of rare variant 475}
 \item{Input_476}{Presence and absense of rare variant 476}
 \item{Input_477}{Presence and absense of rare variant 477}
 \item{Input_478}{Presence and absense of rare variant 478}
 \item{Input_479}{Presence and absense of rare variant 479}
 \item{Input_480}{Presence and absense of rare variant 480}
 \item{Input_481}{Presence and absense of rare variant 481}
 \item{Input_482}{Presence and absense of rare variant 482}
 \item{Input_483}{Presence and absense of rare variant 483}
 \item{Input_484}{Presence and absense of rare variant 484}
 \item{Input_485}{Presence and absense of rare variant 485}
 \item{Input_486}{Presence and absense of rare variant 486}
 \item{Input_487}{Presence and absense of rare variant 487}
 \item{Input_488}{Presence and absense of rare variant 488}
 \item{Input_489}{Presence and absense of rare variant 489}
 \item{Input_490}{Presence and absense of rare variant 490}
 \item{Input_491}{Presence and absense of rare variant 491}
 \item{Input_492}{Presence and absense of rare variant 492}
 \item{Input_493}{Presence and absense of rare variant 493}
 \item{Input_494}{Presence and absense of rare variant 494}
 \item{Input_495}{Presence and absense of rare variant 495}
 \item{Input_496}{Presence and absense of rare variant 496}
 \item{Input_497}{Presence and absense of rare variant 497}
 \item{Input_498}{Presence and absense of rare variant 498}
 \item{Input_499}{Presence and absense of rare variant 499}
 \item{Input_500}{Presence and absense of rare variant 500}
 \item{Output_1}{Disease outcome or phenotype 1}
 \item{Output_2}{Disease outcome or phenotype 2}
 \item{Output_3}{Disease outcome or phenotype 3}
}
}
\usage{
boolean_input_mult_df
}
\description{
A synthetic dataset containing information about 5000
individuals (rows) and 1000 rare variants (columns)
and 3 outcome variables.
}
\keyword{datasets}
