% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_enrichment_depletion.R
\name{compare_enrichment_depletion}
\alias{compare_enrichment_depletion}
\title{Compare the enrichment in combinations of input variables between the binary outcomes (case/control)}
\usage{
compare_enrichment_depletion(boolean_input_df, combo_length, min_indv_threshold,
                             max_freq_threshold, input_format, output_format,
                             pval_filter_threshold, adj_pval_type, min_power_threshold,
                             sample_names_ind)
}
\arguments{
\item{boolean_input_df}{An input Boolean dataframe with multiple input and a single binary outcome variable}

\item{combo_length}{The length of the combinations specified by the user}

\item{min_indv_threshold}{Minimum number of instances that support the combination}

\item{max_freq_threshold}{Maximum fraction of the cohort size that could support a combination (i.e., filter out highly frequent events)}

\item{input_format}{Optional | Naming convention used for input variables (Default = 'Input_')}

\item{output_format}{Optional | Naming convention used for output variables (Default = 'Output_')}

\item{pval_filter_threshold}{Optional | p-value cut-off to use to identify significant combinations in cases (Default = 0.05)}

\item{adj_pval_type}{Optional | Type of multiple testing corrections to use (Default = 'BH'; Alternative option = 'bonferroni')}

\item{min_power_threshold}{Optional | Minimum statistical power (at 5\% sig.threshold) required for significant combinations to be returned in the results (Default = 0.7)}

\item{sample_names_ind}{Optional | Indicator to specify if the output should includes row names that support each significant combination (Default = 'N'; Alternative option = 'Y')}
}
\value{
A dataframe with the list of multiple-testing adjusted statistically significant combinations
        along with quantitative measures (frequencies, p-values etc) that support the findings.
}
\description{
This function takes a Boolean dataframe as input and quantifies the enrichment in the observed frequency of combinations
that meet the criteria specified by the users compared to their corresponding expectation derived under the assumption
of independence between the constituent elements of each combination. The function then reports the multiple-testing
adjusted significant combinations in which enrichment is observed in cases and depletion is observed in controls.
}
\examples{

   compare_enrichment_depletion(boolean_input_df, 3, 5, 0.25, input_format = 'Input_',
                                output_format = 'Output_', adj_pval_type = 'bonferroni',
                                sample_names_ind = 'N')


}
\author{
Vijay Kumar Pounraja
}
