% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input_list}
\alias{input_list}
\title{A list of 50 random input variables}
\format{
A list of 50 random input variables:
}
\usage{
input_list
}
\description{
A list of 50 random input variables
}
\keyword{datasets}
